[Fullstack] Fix destroy of VM fixture
When FakeFullstackMachine is connected to Linuxbridge
instead of Open vSwitch bridge, bridge object don't have
delete_port() method and it is not necessary to remove
"tap" device from host namespace.
This device will be removed together with deletion of
VM namespace.

Closes-Bug: #1735571

+++ neutron/tests/fullstack/resources/machine.py
@@ -198,7 +198,11 @@ class FakeFullstackMachine(machine_fixtures.FakeMachineBase):
             {'port': {pbs.HOST_ID: ''}}
         )
         # All associated vlan interfaces are deleted too
-        self.bridge.delete_port(self.port.name)
+        # If VM is connected to Linuxbridge it hasn't got "delete_port" method
+        # and it is not necessary to delete tap port connected to this bridge.
+        # It is veth pair and will be removed together with VM namespace
+        if hasattr(self.bridge, "delete_port"):
+            self.bridge.delete_port(self.port.name)
 
         ip_lib.delete_network_namespace(self.namespace)
